Bug 1209438 - [Flame KK][Music NGA]User can scroll the music search result page,even list is empty.
