+++ apps/music/css/view.css
@@ -11,6 +11,7 @@ body {
   margin: 0;
   width: 100%;
   height: 100%;
+  overflow: hidden;
   -moz-user-select: none;
 }
 
@@ -63,4 +64,16 @@ gaia-fast-list img {
 gaia-fast-list img ~ h3,
 gaia-fast-list img ~ p {
   padding-right: 60px;
+}
+
+music-search-box {
+  height: 37px;
+}
+
+music-search-results {
+  position: absolute;
+  top: 37px;
+  left: 0;
+  width: 100%;
+  height: calc(100% - 37px);
 }
