+++ apps/music/resources.js
@@ -22,7 +22,8 @@ const RESOURCES = [
   '/elements/music-overlay.js',
   '/elements/music-rating.js',
   '/elements/music-scan-progress.js',
-  '/elements/music-search.js',
+  '/elements/music-search-box.js',
+  '/elements/music-search-results.js',
   '/elements/music-seek-bar.js',
   '/elements/music-tab-bar.js',
   '/elements/music-view-stack.js',
