+++ apps/music/views/artists/view.js
@@ -4,30 +4,32 @@
 var ArtistsView = View.extend(function ArtistsView() {
   View.call(this); // super();
 
-  this.searchBox = document.getElementById('search');
+  this.searchBox = document.getElementById('search-box');
+  this.searchResults = document.getElementById('search-results');
   this.list = document.getElementById('list');
 
-  var searchHeight = this.searchBox.HEIGHT;
-
-  this.searchBox.addEventListener('open', () => window.parent.onSearchOpen());
-  this.searchBox.addEventListener('close', () => {
-    this.list.scrollTop = searchHeight;
-    window.parent.onSearchClose();
-  });
   this.searchBox.addEventListener('search', (evt) => this.search(evt.detail));
-  this.searchBox.addEventListener('resultclick', (evt) => {
+
+  this.searchResults.addEventListener('open', () => {
+    this.client.method('searchOpen');
+  });
+
+  this.searchResults.addEventListener('close', () => {
+    this.client.method('searchClose');
+    this.list.scrollTop = this.searchBox.HEIGHT;
+  });
+
+  this.searchResults.addEventListener('resultclick', (evt) => {
     var link = evt.detail;
     if (link) {
       this.client.method('navigate', link.getAttribute('href'));
     }
   });
 
-  this.searchBox.getItemImageSrc = (item) => {
-    return this.getThumbnail(item.name);
-  };
+  this.searchResults.getItemImageSrc = (item) => this.getThumbnail(item.name);
 
-  this.list.scrollTop = searchHeight;
-  this.list.minScrollHeight = `calc(100% - ${searchHeight}px)`;
+  this.list.scrollTop = this.searchBox.HEIGHT;
+  this.list.minScrollHeight = `calc(100% - ${this.searchBox.HEIGHT}px)`;
 
   this.list.configure({
     getItemImageSrc: (item) => {
@@ -82,6 +84,10 @@ ArtistsView.prototype.getThumbnail = function(filePath) {
 };
 
 ArtistsView.prototype.search = function(query) {
+  if (!query) {
+    return Promise.resolve(this.searchResults.clearResults());
+  }
+
   return document.l10n.formatValue('unknownArtist').then((unknownArtist) => {
     return this.fetch('/api/search/artist/' + query)
       .then(response => response.json())
@@ -96,8 +102,7 @@ ArtistsView.prototype.search = function(query) {
           };
         });
 
-        this.searchBox.setResults(results);
-        return results;
+        return this.searchResults.setResults(results);
       });
   });
 };
