+++ apps/music/views/home/index.html
@@ -23,14 +23,16 @@
   <script defer src="/components/gaia-component/gaia-component.js"></script>
   <script defer src="/components/gaia-fast-list/gaia-fast-list.js"></script>
   <script defer src="/components/gaia-sub-header/gaia-sub-header.js"></script>
-  <script defer src="/elements/music-search.js"></script>
+  <script defer src="/elements/music-search-box.js"></script>
+  <script defer src="/elements/music-search-results.js"></script>
   <script defer src="/shared/js/sanitizer.js"></script>
   <script defer src="/js/config.js"></script>
   <script defer src="/js/view.js"></script>
   <script defer src="view.js"></script>
 </head>
 <body role="application">
-  <music-search id="search"></music-search>
+  <music-search-box id="search-box" results="#search-results"></music-search-box>
   <section id="tiles"></section>
+  <music-search-results id="search-results"></music-search-results>
 </body>
 </html>
