+++ apps/music/views/home/view.css
@@ -1,3 +1,11 @@
+body {
+  overflow-y: auto;
+}
+
+body[data-search="true"] {
+  overflow: hidden;
+}
+
 .tile {
   border-bottom: solid 1px var(--background);
   box-sizing: border-box;
