+++ apps/music/views/home/view.js
@@ -6,13 +6,23 @@ var HomeView = View.extend(function HomeView() {
 
   this.thumbnailCache = {};
 
-  this.searchBox = document.getElementById('search');
+  this.searchBox = document.getElementById('search-box');
+  this.searchResults = document.getElementById('search-results');
   this.tiles = document.getElementById('tiles');
 
-  this.searchBox.addEventListener('open', () => window.parent.onSearchOpen());
-  this.searchBox.addEventListener('close', () => window.parent.onSearchClose());
   this.searchBox.addEventListener('search', (evt) => this.search(evt.detail));
-  this.searchBox.addEventListener('resultclick', (evt) => {
+
+  this.searchResults.addEventListener('open', () => {
+    this.client.method('searchOpen');
+    document.body.dataset.search = true;
+  });
+
+  this.searchResults.addEventListener('close', () => {
+    this.client.method('searchClose');
+    document.body.removeAttribute('data-search');
+  });
+
+  this.searchResults.addEventListener('resultclick', (evt) => {
     var link = evt.detail;
     if (link) {
       if (link.dataset.section === 'songs') {
@@ -23,9 +33,7 @@ var HomeView = View.extend(function HomeView() {
     }
   });
 
-  this.searchBox.getItemImageSrc = (item) => {
-    return this.getThumbnail(item.name);
-  };
+  this.searchResults.getItemImageSrc = (item) => this.getThumbnail(item.name);
 
   this.tiles.addEventListener('click', (evt) => {
     var link = evt.target.closest('a[data-file-path]');
@@ -108,6 +116,10 @@ HomeView.prototype.queueSong = function(filePath) {
 };
 
 HomeView.prototype.search = function(query) {
+  if (!query) {
+    return Promise.resolve(this.searchResults.clearResults());
+  }
+
   var results = [];
 
   return document.l10n.formatValues(
@@ -128,7 +140,7 @@ HomeView.prototype.search = function(query) {
 
         results = results.concat(albumResults);
 
-        this.searchBox.setResults(results);
+        this.searchResults.setResults(results);
         return albumResults;
       });
 
@@ -147,7 +159,7 @@ HomeView.prototype.search = function(query) {
 
         results = results.concat(artistResults);
 
-        this.searchBox.setResults(results);
+        this.searchResults.setResults(results);
         return artistResults;
       });
 
@@ -166,7 +178,7 @@ HomeView.prototype.search = function(query) {
 
         results = results.concat(songResults);
 
-        this.searchBox.setResults(results);
+        this.searchResults.setResults(results);
         return songResults;
       });
