+++ apps/music/views/songs/view.js
@@ -4,18 +4,22 @@
 var SongsView = View.extend(function SongsView() {
   View.call(this); // super();
 
-  this.searchBox = document.getElementById('search');
+  this.searchBox = document.getElementById('search-box');
+  this.searchResults = document.getElementById('search-results');
   this.list = document.getElementById('list');
 
-  var searchHeight = this.searchBox.HEIGHT;
-
-  this.searchBox.addEventListener('open', () => window.parent.onSearchOpen());
-  this.searchBox.addEventListener('close', () => {
-    this.list.scrollTop = searchHeight;
-    window.parent.onSearchClose();
-  });
   this.searchBox.addEventListener('search', (evt) => this.search(evt.detail));
-  this.searchBox.addEventListener('resultclick', (evt) => {
+
+  this.searchResults.addEventListener('open', () => {
+    this.client.method('searchOpen');
+  });
+
+  this.searchResults.addEventListener('close', () => {
+    this.client.method('searchClose');
+    this.list.scrollTop = this.searchBox.HEIGHT;
+  });
+
+  this.searchResults.addEventListener('resultclick', (evt) => {
     var link = evt.detail;
     if (link) {
       this.queueSong(link.dataset.filePath);
@@ -24,10 +28,10 @@ var SongsView = View.extend(function SongsView() {
     }
   });
 
-  this.searchBox.getItemImageSrc = (item) => this.getThumbnail(item.name);
+  this.searchResults.getItemImageSrc = (item) => this.getThumbnail(item.name);
 
-  this.list.scrollTop = searchHeight;
-  this.list.minScrollHeight = `calc(100% - ${searchHeight}px)`;
+  this.list.scrollTop = this.searchBox.HEIGHT;
+  this.list.minScrollHeight = `calc(100% - ${this.searchBox.HEIGHT}px)`;
 
   this.list.configure({
     getItemImageSrc: (item) => {
@@ -93,6 +97,10 @@ SongsView.prototype.getThumbnail = function(filePath) {
 };
 
 SongsView.prototype.search = function(query) {
+  if (!query) {
+    return Promise.resolve(this.searchResults.clearResults());
+  }
+
   return document.l10n.formatValues(
     'unknownTitle', 'unknownArtist'
   ).then(([unknownTitle, unknownArtist]) => {
@@ -109,8 +117,7 @@ SongsView.prototype.search = function(query) {
           };
         });
 
-        this.searchBox.setResults(results);
-        return results;
+        return this.searchResults.setResults(results);
       });
   });
 };
